requirements.txt
streamlit==1.28.1
azure-cognitiveservices-speech==1.34.0
pydub==0.25.1
pandas==1.5.3
streamlit-audiorecorder
openai==1.3.0
jaconv==0.3.4
cutlet==0.1.19
pykakasi==2.2.1
streamlit-audiorecorder
streamlit-webrtc==0.47.1
requests==2.31.0